requirements.txt
pandas~=2.2.3
numpy~=2.1.3
requests~=2.32.3
selenium~=4.27.1
python-dotenv~=1.0.1
openpyxl